ResizingArrayStack
Create a class called ResizingArrayStack, which uses a resizing-array to store items.

Your stack must store any type.

Specification

Constructor: initializes array at size 1.
push method: accepts any type and adds it to the top of the stack (returns void).
When a push would cause the array to overflow, double the size of the array.
Must throw an IllegalArgumentException if a null item is given.
pop method: removes and returns the item at the top of the stack (accepts nothing).
When a pop would cause the array to reduce to one fourth its current capacity, halve the size of the array.
If the stack is empty then this method must throw an IllegalStateException.
peek method: returns, but does not remove, the top item from the stack (accepts nothing).
If the stack is empty then this method must throw an IllegalStateException.
isEmpty method: tests to see if the stack is empty.
size method: returns the number of items in the stack.
Your implementation must avoid loitering. Loitering is holding a reference to an object when it is no longer needed. The Java garbage collector can reclaim memory for an object only if there are no outstanding references to it. So eliminating these references (by setting them to null) saves memory.